The ProLint Webserver v2.0
ProLint is a platform to analyze and visualize lipid-protein
interactions. It includes several tools: a python library, a C-written
g_surf program, an online free to use webserver, and this open-source
project which allows you to have the ProLint webserver running locally
on your machine with just a few clicks using Docker.